Add FlatMap
This is similar to `Map`, but instead of taking a function for a `T` to a `U` and returning a `Domain<U>`, it takes a function from a `T` to a `Domain<U>` and returns a `Domain<U>`.

This allows users to create domains which depend on the output of other domains. Type-theory-wise, this implements the monadic "bind" operation for `Domain`.

PiperOrigin-RevId: 474667143