Cache bazel results across CI runs.
We rely on actions/cache to save ~/.cache/bazel. Note that caches don't get
automatically updated, they get only re-created when there's a cache miss.

We can add a workflow later to clear the caches before the scheduled nightly run
to create a fresh cache for the day.

I was able to re-run our current ~15 min CI run in 1.5 min, a 10x speed-up.

PiperOrigin-RevId: 474911646