Schedule nightly Bazel cache reset.
I am using the scheduled nightly run that was previously used for running the CI
against a recent version of Bazel. Instead of either the recent version or the
pinned version, we now simply use the default version from ubuntu-latest. This
will be stable enough to avoid potential issues where the cache would be
invalidated in the regular runs due to version mismatch.

PiperOrigin-RevId: 475402174