Adds the ability to update optional fields without changing the nullness policy.
Before this CL, we decompose a field customization into the base domain customization and rule
(optional/repeated/required) customization and store them as policies. Then, we rebuilt the
custom domain using separate policies. This is problematic if the use wants to use their
own implementation for optional and vector domains. To avoid this, we store optional and
repeated in their entirety (without decomposition).

In addition, when user uses With<Type>Field, we don't use any default implementation for the
rule (optional/repeated/required). The implementation will be decided by the policy.

PiperOrigin-RevId: 477000218