FuzzTest: make use of SA_ONSTACK flag for signal handlers
For programs with Go threads present, signal handlers must be initialized with
the `SA_ONSTACK` flag due to Go threads inherently having small stacks [0]. This
flag ensures that a dedicated stack is used during the execution of a signal
handler. Failing to initialize signal handlers with `SA_ONSTACK` results in the
abnormal termination of the program in the instance that a signal is received.

FTR, we're making use of FuzzTest to test both our C++ and Go code. We use `cgo`
as a mechanism for exposing Go functions to C++ test code and hence Go threads
being present in the C++ test binary.

Apart from an increased memory footprint due to the kernel having to install a
dedicated stack region for the signal handler to run in, I don't immediately
foresee there being any other problems using the SA_ONSTACK by default.

[0] https://pkg.go.dev/os/signal#hdr-Go_programs_that_use_cgo_or_SWIG

PiperOrigin-RevId: 490049359