Fixes map field issues with required fields.
There are two issues with the required fields, which are addressed here:
(1) Map field values could be messages with required fields. We treat map field values that are messages as required. This way, we make sure the `value` is initialized while the message itself could be empty.
(2) There could be required fields next to map fields. To address this, we fix serialization-deserialization of map fields to remove duplicates. `corpus_copy` contains only one field (a map field) which goes through a mutation. So, the serialization should be done partially and we should not be worried about other required fields.

PiperOrigin-RevId: 499842705