Refactor type erasure for Domain/FixtureDriver/FuzzerImpl to reduce compile time costs.
Before this change every `FUZZ_TEST` got its own copy of the runtime because it was a template based on the fixture (which was a unique type) and based on the domain. Also, because it is templates it was being compiled in the user's TU which is instrumented for coverage.
After this change there is a single runtime hidden in an uninstrumented .cc file. The driver is still a template but it is reduced to only require the signature of the callable and the signature of the domain.

PiperOrigin-RevId: 501345540