Replace type-parameterized PRNG& with absl::BitGenRef.
The type-parameterized approach has not shown to be useful. Also, meanwhile
we've started using absl::BitGenRef at the top level, which makes the templates
get instantiated as absl::BitGenRef&. This is contrary to the recommendation of
passing absl::BitGenRef by value.

To rectify this, and generally improve code readability and consistency, this CL
changes all occurrences of PRNG& with absl::BitGenRef.

PiperOrigin-RevId: 501903888