Adds support for customizing nullness oneof fields.
An oneof (e.g., `my_oneof`) is "AlwaysSet" iff (a) all its fields are
"AlwaysSet" or (b) `.WithOneofAlwaysSet("my_oneof")` is called. In addition
any field in the oneof that is "Unset" won't be set in any situation.

To implement this, we also avoid selecting "Unset" oneof fields for mutation.

PiperOrigin-RevId: 503217776