Implement seeded domains by adding `.WithSeeds()` to domains.
Seeded domains occasionally return one of the seeds as a result of `.Init()`.

NOTE: The seeds are accepted as user values and converted to corpus values using
`.FromValue()`. As a consequence, setting the seeds won't work for domains for
which `.FromValue()` fails (e.g., `MapOf`, `FlatMap`, and anything derived from
them like `Arbitrary<absl::Duration>`). Moreover, currently `.FromValue()`
doesn't validate the inputs, so it is possible to store a seed that doesn't
meet the domain's constraints (e.g., `InRange(0, 10).WithSeeds({42})` works).
This will be fixed separately.

The goal of the added unit tets is to almost always succeed when the
functionality is implemented and almost always fail when the functionality is
missing. For most domains this usually means that it is sufficient to check that
one or two seeds appear at least once among 1000 generated values, since it is
unlikely that those seeds would have been generated otherwise. But this is not
sufficient for domains with few elements. For example, for `Arbitrary<bool>` we
are surely going to see `true` generated at least once regardless of whether it
is a seed or not. For such domains I derived lower bounds for the number of seed
appearances using the [Hoeffding's inequality][1] so that the tests are robuts.

[1]: https://en.wikipedia.org/wiki/Binomial_distribution#Tail_bounds

PiperOrigin-RevId: 519828975