Add an end-to-end test for seeded domains.
I also improved the manual dictionary test, which was not effective before.
The test was checking that we can hit a reversal of a string, which supposedly
wouldn't work without a manual dictionary. Turns out it works even without it.
The compiler simply optimized away the reversal.

Now we're using a more complicated permutation of the string.

PiperOrigin-RevId: 523214244