Extend the `.WithSeeds()` API to accept seed providers.
The extended API can be used to register a function that returns seeds. This is
useful as when the seeds need to be initialized dynamically, e.g., to resolve
the static initialization order fiasco.

The seed provider can be a free function or a function member on a fixture, the
latter being useful when the seeds also depend on the fixture's state.

PiperOrigin-RevId: 524956097