Implement seed providers for seeded domains.
This is analogous to `.WithSeeds()` for test registration: we allow passing
a function that generates seeds lazily, so that the users can avoid static
initialization issues. Unlike `.WithSeeds()` for test registration, here the
seed provider can only be a free function and not a fixture member, since
at the level of domains we have no knowledge of test fixtures.

I also implemented seed validation for seeded domains.

A few notes:

* To make things consistent with the registration-level `.WithSeeds()`, the
  domain-level `.WithSeeds()` now appends to the previously added seeds.

* If the user values are bools, then the seed provider returns
  `std::vector<bool>`, which cannot be converted to `absl::Span<const bool>`.
  To avoid complications, the previous `.WithSeeds()` overload now takes
  `const std::vector<ValueType>&` as input.

* There is a weird corner case of `.WithSeeds({0})`: `{0}` can be used to
  initialize `std::function`, which results in an ambiguous call.

* There was a bug in `FromValue` for `OneOf` that was uncovered by one of the
  `InitGeneratesSeeds` tests once I added validation, so I fixed that as well.

PiperOrigin-RevId: 531605002