#Centipede Do not overwrite telemetry files in purely distilling jobs
Do not overwrite the telemetry files (corpus stats, coverage report, rusage) if this is a purely distilling job (--num_runs==0). The post-distillation `*.distilled.*` versions of the files are still written. This enables on-the-fly distillation of a corpus in-progress without stopping the fuzzing job that continues to update it and writing the distilled corpus files to the same `--workdir`.

PiperOrigin-RevId: 532308533