#Centipede Ignore short/long cmp entries instead of skipping the rest of the entries.
This is a saner choice to deal with short/long entries without losing valuable information after them.

As a drive-by, enhanced the tests for FuzzTest mutator a bit.

Note that the tests don't touch the case of short entries because even if the entries are skipped, the same effect of inserting/overwriting with short entries can still easily manifest from inserting/overwriting random bytes.

PiperOrigin-RevId: 542978215