#Centipede Add new target API to use the execution metadata to the custom mutator in the mutation process.
To access the metadata, the custom mutator would have to implement the new Centipede callback, which would be called by the Centipede runner for each mutation batch.

Note that we cannot directly pass the ExecutionMetadata class because the target interface is in plain C - some data point to consider on improving/revamping the target interface.

PiperOrigin-RevId: 544368701