#Centipede Document and tolerate the case where custom mutator returns empty set of mutants.
When such case happens it would fall back to use the internal mutator (which should be rare but okay even if it happens frequently).

With the empty result tolerated, we no longer need to check it in the mutator detection. This helps with the reliability of custom mutator detection, as some mutators (e.g. the one in test_fuzz_target) may refuse to mutate by chance.

PiperOrigin-RevId: 556090144