Make sure WithFields[AlwaysSet/Unset] overwrite a With[Min/Max]RepeatedFieldsSize.
Currently, `.WithFieldsUnset().WithMinRepeatedFieldsSize(1)` yields an error because the first method sets the maximum size to 0 and the later sets the minimum size to 1. Similarly, `WithMaxRepeatedFieldsSize(0).WithFieldsAlwaysSet()` yields an error. To address the issue, make sure `WithFields[AlwaysSet/Unset]` methods will overwrite `With[Min/Max]RepeatedFieldsSize` methods.

One of the consequences which is a little unexpected is that in the following, the maximum size will set to default and not to the user-provided value:

.WithMaxRepeatedFieldsSize(7).WithFieldsAlwaysSet()

PiperOrigin-RevId: 561642709