#Centipede Allow disabling user feature domains from command line
For example, --user_feature_domain_mask=0xfffe will disable usr0 / the first user feature domain.

Unlike existing flags that control if a feature is generated, this flag causes the feature to be discarded from the feature vectors that centipede processes. This means that if user_feature_mask becomes more restrictive between Centipede runs, features in the existing corpus may be filtered out as the corpus is loaded and in turn inputs previously deemed interesting enough to put in the corpus may be deemed uninteresting when they are reloaded.

This filtering will also affect distillation. --user_feature_domain_mask can be used to distill a corpus as if certain user feature domains did not exist.

PiperOrigin-RevId: 563590617