#Centipede Update exit status decoding so that it also works correctly when running in Docker container.
The exit codes returned (in both fork and non-fork modes) differ across environments. The current code pattern using the `wait` macros is the recommended way for decoding them in a platform agnostic way. However, it doesn't seem to work correctly at least in one case: in a docker container, in non-fork mode for a process that is terminated by a signal.

There is no clear documentation that I have been able to find out, but the following has been observed from the running behavior. When run in a Docker container, the exit code resulting from a `system()` call is an encoding of the exit code returned from the argument process. This encoding embeds it in the 2nd lowest byte (rather than the lowest byte). Consequently, the resulting exit code is incorrectly determined (by the `WIFEXITED` macro) as normal termination even if the underlying process was terminated by a signal. In these cases, the exit status of interest (of the invoked process) has to be obtained through one level of decoding (through the `WEXITSTATUS` macro) before further processing/decoding is done.

Exit codes under normal termination are less than 128, with exit codes greater than 128 being used to encode termination by a signal `n` as `128+n` in some shells (e.g., see https://www.gnu.org/software/bash/manual/bash.html#Exit-Status, https://tldp.org/LDP/abs/html/exitcodes.html). It is therefore safe (across different environments) to consider a decoded exit code of more than 128 to be inconsistent with normal termination and apply this one layer of decoding in this case. This is what is done in the current CL.

PiperOrigin-RevId: 568926176