#Centipede Refactor AnalyzeCorpora to read BinaryInfo from disk.
This changes the behavior to read all of the corpus shards within the workdirs
passed to --analyze. Previously, this would use the --total_shards flag to
determine how many shards should be read. By reading all of the corpus shards
within the workdir we avoid coupling the analyze_corpora.{cc,h} library with
the centipede flags. This will enable the analyze corpora functionality to be
used within a library call from a non-centipede binary.

PiperOrigin-RevId: 570665467