#Centipede Support memory sanitizer and add new tests.
MSAN tests are not enabled in OSS since it requires the standard libraries to be instrumented too.

Specially, no_sanitize("memory") is applied on a few new functions to avoid false positives. Also renamed the dtor of the global runner state since MSAN would poison the state memory and crash on further access from the watchdog.

PiperOrigin-RevId: 577888953