#Centipede Allow HashOfFileContents to use a remote file.
`ReadFromLocalFile` (used in the old implementation of `HashOfFileContents`)
returns early without setting any contents when the file read fails. This
change somewhat adjusts the functionality by returning an empty hash if the
path is empty. This new version will CHECK fail more frequently than the old
one, e.g., if the path is non-empty but the file doesn't exist.

PiperOrigin-RevId: 577952195