#Centipede Handling non-existent files in RiegeliReader/Writer.
If a Riegeli::RecordReader/Writer object is attempted to be opened with a non-existent file or one that cannot be created respectively, although it will fail, it leaves the object in an opened state and a subsequent close will fail as well because of the non-existent file. These changes address this situation more uniformly through RiegeliReader/Writer as well as make the Open/Close symmetric for Reader/Writer.

PiperOrigin-RevId: 583218675