#Centipede RUsageProfiler: Print final report with source location set to client code
The previous version would print all snapshots and such as if the client's code itself printed them by using `LOG(...).AtLocation(client_file, client_line)`, and even print the final report's title the same way. This is intentional: it makes printed metrics easily attributable to the specific line in the client code that requested the measurement. However, the report's body was printed on behalf of the profiler itself, making it more difficult to use (e.g. when filtering the log by the client source name).

PiperOrigin-RevId: 584450583