Create content-aware mutator that understands instructions
This mutator tries to keep direct branches pointing inside the program and pointing at valid instruction boundaries. Previously, direct branches were underrepresented in Silifuzz tests because with a high probability a random branch would point outside the test.

In order to fix up direct branches, however, we need to be able to decompile random programs. This is a bit tricky on x86_64, however, because branches can jump inside other instructions. The mutator boldly ignores this situation, and for the most part it seems to work. As long as the mutator fixes up direct branches, the only way tests can jump into the middle of instructions is with an indirect branch - but currently those are even harder to discover than direct branches. So the mutator may be suboptimal for cases where the test jumps into the middle of an instruction, but those cases should be rare.

PiperOrigin-RevId: 590326337