Adding support for the InGrammar domain to allow inserting a whitespace between subrules/tokens of a production rule.
Specifically, we added the "insert_space_between_blocks" flag for the codegen binary, as well as "insert_whitespace" flag in the BUILD rule. The flags are FALSE by default. If set to true, the input ANTLR grammar's parser production rules will be modified to insert a single literal space between blocks.

For instance, a parser rule with two alternatives like below

Rule : A | B C;

will be modified to become

Rule: A | B ' ' C;

The first alternative only has a single subrule, hence no space is appended at the end. The second alternative has two subrules, hence a space is inserted in between.

For context, in ANTLR, many grammars skip whitespaces by using the "->skip" or "->channel(HIDDEN)" in the lexer rules for whitespaces. In that case, it is sometimes desirable for the InGrammar domain to generate strings that does NOT skip whitespaces. This is helpful for the lexer/parser in the fuzzing target to disambiguate certain tokens, like keywords v.s. identifiers.

PiperOrigin-RevId: 605037377