#Centipede Fuzzing stats: append, not overwrite, CSV report whenever possible
If a fuzzing stats CSV file already exists and has the same CSV header as the new one, then append to that file rather than overwriting it. If the new CSV header is different, then make a backup copy of the old file and overwrite the file.

This is to preserve the record of the old progress when Centipede is restarted in the same workdir for whatever reason.

PiperOrigin-RevId: 605135239