Encapsulate the type erasure in the domain API.
We split the DomainBase class into DomainModel and DomainBase. The former
implements the delegation that is part of the type erasure, and the latter
is left as the root of the domain hierarchy. This way the implementation of
the type erasure abstraction (UntypedDomainConcept, TypedDomainConcept) is
not leaked in the concrete domain implementations.

Furthermore, in the runtime we now use the domains through a new UntypedDomain
interface. Again, this is so that we don't leak the implementation of the type
erasure abstraction.

One non-trivial change that was necessary to achieve the encapsulation was to
get rid of a hack for printing counterexamples and reproducer tests in the
runtime. Previosly, this was done by calling a special method called
`UntypedPrintCorpusValue`, which was overriden in `AggregateOfImpl` to allow
printing individual elements of an aggregate value. This logic is now moved
to `AggregatePrinter`, which provides a way to customize formatting of
aggreagate values so that it can print counterexamples and reproducers directly.
The solution is not ideal (the printer API is due for a cleanup of its own),
but it is cleaner than before and gets the job done.

PiperOrigin-RevId: 609804673