#Centipede Eliminate RSS spikes when dumping corpus stats & coverage reports
The contents of both files was entirely materialized in memory first (as `std::ostringstream`), then written to the destination on disk. For some corpora, those contents took tens of GBs (over 100 GB has been observed), causing spikes in RAM usage, in particular RSS.

Changed the signature of the respective APIs to write directly to a file, with buffering. This eliminates RSS spikes while keeping the latency at approximately the same level.

PiperOrigin-RevId: 622985175