Expose API to parse known absl flags.
This change makes it easier for consumers such as Chromium to use fuzztest. It
changes argument handling in two ways:

1) It exposes a new 'InitAbslFlagsForFuzztest' API. Chromium's test suites
   otherwise needed to call into absl flags directly to initialize absl's
   command line; this was problematic in projects where absl flags are not
   normally used. Integration of fuzztest is easier if we need to call
   only fuzztest APIs instead of absl flags directly.

   A key point of this new API is that it does not call
   absl::ParseCommandLine, but instead absl::ParseAbseilFlagsOnly.
   This ignores unknown command-line options, which is important in
   environments where other code may use non-absl parsing of command lines.

2) fuzztest_gtest_main is amended to use the above function instead
   of absl::ParseCommandLine. This is because, when used with libfuzzer
   or other non-centipede fuzzing engines, those fuzzing engines themselves
   absorb command linm arguments, and thus the previous use of
   absl::ParseCommandLine always failed unless a complex
   '-undefok=' argument was constructed listing all the possible options
   that could be passed to libfuzzer.

As an indication of the complexity here, in a typical Chromium test,
command-line arguments may flow to four places:
* FuzzTest's absl flags
* libfuzzer, passed there via InitFuzzTest (or some other fuzzing engine)
* Chromium's own test frameworks (flags like --single-process-tests)
* Chromium's underlying code, e.g. for enabling runtime features
  or disabling the sandbox.

Because of the many areas of code handling command line arguments, it's not
practical to declare them all in advance either using absl flags, nor using
`-undefok` and that's why it's important that FuzzTest uses
absl::ParseAbseilFlagsOnly to avoid erroring in the face of unknown extra
command line arguments.

This CL does not fix a third problem in this area, which is that
the fuzztest engine stores the given argc and argv and assumes it never
subsequently changes - this assumption is wrong in projects such as
Chromium where other layers of software may absorb and modify the
command line later, and results in use-after-return bugs. However
this can be easily worked around by storing an extra copy of the original
argc and argv for use by fuzztest, disregarding subsequent changes.

PiperOrigin-RevId: 625785463