Use ParseTextProtoOrDie instead of ParseTestProto in suggested reproducer tests
ParseTestProto potentially hides its failure to parse the proto, leading to bugs in test code.

This CL replaces it with ParseTextProtoOrDie, which crashes loudly on invalid input. The generated reproducer tests always pass a string literal to proto parsing, so if it does not crash on presubmit, such a test is safe. The CL also fully qualifies the function name to help make sure that the result compiles.

PiperOrigin-RevId: 625797971