#Centipede Add `ClearEarlyExitRequest()` and call it from `CentipedeMain()`.
When Centipede is used as a library, `CentipedeMain()` may be called multiple
times. This is the case with FuzzTest when it executes multiple fuzz tests. In
those cases we need to clear the request to exit early, otherwise all tests
except for the first one simply exit as the flag remains set from the first
test.

I also tweaked the implementation a bit:

-  I grouped the two variables so that they are updated in a single atomic step.

-  I relaxed the memory accesses to release/acquire.

PiperOrigin-RevId: 626486868