#Centipede Extend the remote file API and fix crash reporting.
For the remote file API:

- Added `ByteArray` overloads for `RemoteFile{Get|Set}Contents`.
- `RemoteMkdir` creates missing parent directories.
- `RemoteListFiles` now supports both direct listing and recursive expansion.
- Added `RemoteRename` and `RemoteDelete`.

As for the crash reporting, the crash inputs were being truncated to 16 bytes
and stored in a "printable" format, with the non-printable characters stored
in the hex (\xXX) format. This was a regression from an earlier change. Now we
again store the full crash inputs in the raw form.

PiperOrigin-RevId: 629089994