Provide a consistent interface for fuzz flags and replay_corpus flags.
It also includes a control flag: whether the time budget is per test or for per test binary run.

Notice that `use_coverage_inputs_` is omitted from the CorpusDatabase. Because it has no effect when the corpus database path is not set. When the corpus database path is set, the coverage inputs are used for seeding in fuzzing mode. In unit test mode, the coverage inputs will be used only when replay_coverage flags are set. As a result, there's no need for `use_coverage_inputs_`.

PiperOrigin-RevId: 660383205