Implement dumping the crash metadata to a path specified in an env variable.
Currently, we only dump the crash type, which is either set by the runtime
(GoogleTest assertion failure, a crashing signal), or taken from a sanitizer
report (e.g., heap-buffer-overflow).

In the future, we will also dump the stack trace (or part of it).

We will also remove the `crash_handler_hook` and replace it with adding another
crash metadata listener. This is the reason for the listener mechanism.

PiperOrigin-RevId: 672962194