Fix a use-after-move bug in `StatsCsvFileAppender`.
Since the class closes the files from the `files_` map in its destructor, it
needs to make sure that it actually owns those files. In particular, it
shouldn't attempt to close the files after a move. We wrap the map in a
`std::unique_ptr`, which becomes `nullptr` after the move.

Additional small fix is to set `curr_file_` to `nullptr` after the move.

PiperOrigin-RevId: 675713637