Ensure that the Centipede runner writes the failure atomically and only once.
The atomicity is needed because the function `WriteFailureDescription()` may
be called from multiple threads: from the main thread, e.g. for a GoogleTest
assertion failure, and from the watchdog thread, e.g. for an OOM.

The function should write the failure only once because for failures like stack
overflow, it gets called twice: the first time for the stack overflow itself,
and the second time from the SIGABRT handler handling the abort that follows the
detected stack overflow.

PiperOrigin-RevId: 695434225