Print mutator log when runner log requested.
In developing fuzzers, it's often useful to print logging from the fuzzer
binary, and centipede provided a command-line argument to do that. However this
argument didn't apply to invocations for the purpose of gathering mutants, thus
it was hard to debug problems when developing custom implementations of
LLVMFuzzerCustomMutator and similar. Add logging in this case too.

PiperOrigin-RevId: 701867237