Use `std::vector` instead of the custom `InplaceVector` for LLVMFuzzerMutate.
The custom `InplaceVector` has a few issues: Its default constructor (with nullptr backend data) and copy behavior (without copying the underlying data) can be confusing and problematic to the domain implementation; The current container domain does not support dictionary-based mutation on `InplaceVector`. Using the standard `std::vector` (with a bit of copying overhead) seems to be a more economical solution than fixing the aforementioned issues.

PiperOrigin-RevId: 704349332