Check time budget in the corpus database update only if time per test is finite.
Combined with the check that makes sure the time per test is finite when we're
fuzzing, we get the following behavior:

- When fuzzing, the time limit per test must be finite and we perform the check.
- When replaying, the time limit may be infinite, and then we don't perform
  the check.
- We still leave the possibility to set a finite time limit even when replaying,
  in which case we perform the check.

The e2e test doesn't properly check this because there the Bazel test timeout is
not set for the child process. I will follow up with a test once I figure out
how to properly test this.

#Centipede

PiperOrigin-RevId: 705923559