Fix non-determinism in `UniqueElementsContainerOf` and `UniqueElementsVectorOf`.
These domains are typically used with sequence containers like `std::vector`,
where the ordering of elements is important. However, by basing the domains on
`absl::flat_hash_set`, the ordering gets lost, leading to non-determinism. This
can lead to unexpected behavior, especially in combination with combinators like
`FlatMap`.

The fix exploits the fact that `AssociativeContainerOfImpl<absl::flat_hash_set>`
uses `std::list` -- a sequence container -- as a corpus type, so it already
preserves the ordering of elements. Furthermore, it automatically dedups
elements while doing `Init` and `Mutate`. The only problematic part are the
functions `FromValue` and `GetValue` in `UniqueElementsContainerImpl`, which
convert the values via an intermediate set and can thus lose the ordering.
The fix is to convert the values directly, element by element.

I also improved validation, which now makes sure that the corpus value indeed
has unique elements.

PiperOrigin-RevId: 715922985