Add execution_id for resuming/skipping test independently.
This is done by checking and writing base_workdir/test.execution_id with the execution_id.

Given a test to run, we check whether the coverage dir (inside the workdir) exists and if the .execution_id stamp matches the current execution ID, then resume the progress if:

  - execution ID mismatch: the test was not run in this execution - create the coverage dir, run for the whole duration, and remove the coverage dir once finished.
  - execution ID matches but the coverage dir is missing: the test was done previously for this exeuction - skip running the test.
  - execution ID matches but the coverage dir exists: the test was run previously but interrupted - run for the duration based on the previously recorded running time in the workdir.

One side-effect is that now the base `workdir.[shard]` directory is not cleaned up upon the completion. Instead, the workdirs inside are removed, but the .execution_id files there kept as tombstones and read by later executions.

PiperOrigin-RevId: 721566201