Crash Centipede on setup failure and prune old crash inputs with setup failures.
Note: We currently don't crash on setup failures that involve exceeding resource
limits, e.g., timeouts or OOMs. Doing that requires connecting the Centipede
watchdog with the FuzzTest runtime to be able to pass information about the
crash type (currently the watchdog directly dumps the crash type to a file and
has no way of knowing whether we're in the test setup or executing an input).
PiperOrigin-RevId: 721812896