Distinguish between undetected and failing custom mutator (fail on the latter).
I extended the IPC protocol for mutation via external binary: the binary now
first writes a marker to indicate whether it has a custom mutator. If it does,
it proceeds by writing mutants, otherwise it successfully exits.

On the Centipede side, this allows distinguishing between an undetected custom
mutator (indicated by the marker) and a failure (indicated by a failed
execution).

The protocol is implemented using a new type of runner result, `MutationResult`,
similarly to the `ExecutionResult` and `BatchResult`.

In addition, I cleaned up the `Mutate()` API so that the results are passed as
return values instead of using output parameters. This results in cleaner, less
error-prone, and more efficient code (e.g., the number of mallocs in the
malloc-counting test went down significantly, so I updated the expected count
there as well).

PiperOrigin-RevId: 731361992