Report skipped test in the runner mode instead of the controller mode.
With this change, FuzzTest no longer needs to instantiate the fixture in the controller mode,
and setup failures in the fixture instantiation can be captured (otherwise they would crash the controller process).
Because of this, a few fixture test expectation become trickier.

Note that there can be setup failures that cannot be captured, such as environment setup and the static SetUpTestSuite function. But this is still a step forward.

PiperOrigin-RevId: 745598295