Remove destabilization of the random number generator in `GetRandomValue()`.
The motivation for destabilization was to prevent users from relying on seed
stability in their tests, as this can lead to maintenance problems for FuzzTest
itself: a "random" test that deterministically succeeds can start
deterministically failing simply due to a change in the domain implementations,
thus blocking us from submitting the change before we fix the user's test.

However, the destabilization does prevent legitimate use cases when the user
wants to rely precisely on the stability guarantees of their generators, and
they make sure they are not using fixed seeds.

We've decided to shift the tradeoff between minimizing maintenance risk and
allowing such use cases in favor of the latter.

PiperOrigin-RevId: 747555423