#Centipede Propagate ExecutionMetadata from shared memory to ByteArrayMutator
When a custom mutator is defined, we call `CentipedeCallbacks::MutateViaExternalBinary` which in turn sends a mutation request to the binary.

The binary (running in a child process) performs the mutation using the inputs passed thru shared memory. This is done in `MutateInputsFromShmem` which invokes the custom mutation code.

This custom mutation code can make use of the `LLVMFuzzerMutate` which is implemented by `CentipedeLLVMFuzzerMutateCallback`.

`CentipedeLLVMFuzzerMutateCallback` uses the `state.byte_array_mutator` without making use of the ExecutionMetadata.
This change propagates the first ExecutionMetadata from shared memory to the byte_array_mutator.

This should improve Centipede's performance with libprotobuf-mutator fuzzers.

PiperOrigin-RevId: 750296827