Support running Centipede in separate processes.
This is enabled when --centipede_bianry_path is set by either fuzzing infrastructure or some special library (not in this CL) that provides the flag during init (so users can opt-in by themselves).

The Centipede process is run using `popen` so that the output can be piped on the fly, while the process will (eventually) recieve SIGHUP and terminate when the parent test process ends (e.g. due to user interruption).

In order to pass serialized configurations in command flags, changed the flag to take base64url encoded string.

PiperOrigin-RevId: 753204996