Export and report crash reproducers from Centipede.
Previously when running with Centipede, crash reproducers are saved only when updating the corpus database, otherwise they are never reported to users.
Reporting reproducers (with reproduction instructions) has been desired in multiple scenarios, but the feature was only available in the legacy engine.
This change implements the feature in FuzzTest/Centipede properly by letting the controller handle the reporting.

One observable difference here with the legacy engine is that FuzzTest/Centipede would find and report multiple reproducers in one run.
In order to avoid repeatedly reporting similar reproduction instructions, the controller would report the reproducers together at the end of the run, with first lines listing the reproducer IDs and metadata (crash types), followed by a reproduction instruction with a placeholder reproducer path. Users can replace the placeholder with the reported IDs to reproduce actual crashes.

PiperOrigin-RevId: 755555620