Fix the corpus value validation for `InRegexp`.
The function `RegexpDFA::DFAPathToString()` that is used in the corpus value
validation wasn't checking that the DFA path is consistent with the DFA,
i.e., that the states along the path are really the ones appearing in the DFA
if we follow the path. It also wasn't checking that the path reaches the end
state.

I also cleaned up the code a bit and added a few general improvements.

PiperOrigin-RevId: 756386731